cleaning

+++ src/main/java/com/example/shoplist/model/Product.java
@@ -8,7 +8,7 @@ public class Product {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "is_bought")
@@ -18,18 +18,16 @@ public class Product {
 
     }
     public Product(String name) {
-     //   this.id = id;
-
         this.name = name;
         isBought = false;
     }
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
